requirements.txt
# Web Framework
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0

# Production WSGI Server
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.10

# HTTP Requests
requests==2.32.4
urllib3==2.5.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10

# Web Scraping
beautifulsoup4==4.13.4
soupsieve==2.7

# Browser Automation (for squad value scraping)
playwright==1.54.0
pyee==13.0.0
greenlet==3.2.3
typing_extensions==4.14.1